+++ flake.nix
@@ -32,7 +32,7 @@
       {
         packages.default = rustPlatform.buildRustPackage {
           pname = "gregorio-lsp";
-          version = "0.9.3";
+          version = "0.9.4";
           src = ./.;
 
           cargoLock = {